text_label_backend/requirements.txt
absl-py==1.4.0
annotated-types==0.6.0
asgiref==3.7.2
astroid==3.2.3
black==24.4.2
blis==0.7.11
boolean.py==4.0
CacheControl==0.13.1
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cyclonedx-python-lib==5.2.0
cymem==2.0.8
defusedxml==0.7.1
dill==0.3.8
Django==3.2.20
django-cleanup==8.0.0
django-cors-headers==4.1.0
django-filter==23.2
django-rest-authtoken==2.1.4
django-zen-queries==2.1.0
djangorestframework==3.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.12.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
html5lib==1.1
idna==3.4
pip-install==1.3.5
isort==5.13.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
license-expression==30.2.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
munkres==1.1.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.25.0
oauthlib==3.2.2
packageurl-python==0.13.1
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
Pillow==8.4.0
pip-api==0.0.30
pip-requirements-parser==32.0.1
pip_audit==2.6.2
platformdirs==4.2.2
preshed==3.0.9
protobuf==4.24.0
py-serializable==0.15.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pylint==3.2.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.1
six==1.16.0
smart-open==6.4.0
snorkel==0.9.9
sortedcontainers==2.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
textblob==0.18.0.post0
the-social-network==0.0.7
thinc==8.2.2
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.0
torch==2.0.1
tqdm==4.66.1
tweepy==4.14.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
weasel==0.3.4
webencodings==0.5.1
Werkzeug==2.3.7
